Web Animations: Fix updating of Players which come back into effect
This patch fixes the case where a player stops being current or in
effect, but later becomes current or in effect again. The current
behaviour has a list of current/in effect players on the timeline,
from which players can only get removed. The new behaviour will be
to maintain a set of players which are current, in effect, or have
had their timing updated.

BUG=334937

Review URL: https://codereview.chromium.org/168983009

git-svn-id: svn://svn.chromium.org/blink/trunk@167484 bbb929c8-8fbe-4397-9dbb-9b2b20218538