Return the right top position for text selections when there are no previous root inline boxes.
The code in RootInlineBox::selectionTop() was doing some extra calculations based on border and
padding to determine the top position of a text selection for a RootInlineBox when there is no
previous box to consider, instead of just returning the m_lineTop value (which is similar to
what is correctly being done already now in RootInlineBox::selectionBottom()).

This caused that the top position for a selection was usually miscalculated, returning to high
values when the line-height value is too big, and too small values when line-height is too
small, causing rendering issues both when drawing cursors and selection rectangles.

BUG=313593
TEST=Try selecting text over lines with too big and too small values for line-height.
R=eae@chromium.org

Review URL: https://codereview.chromium.org/141683003

git-svn-id: svn://svn.chromium.org/blink/trunk@167813 bbb929c8-8fbe-4397-9dbb-9b2b20218538