Clean up compositing reasons definitions for better maintainability
This is a much-needed updating of the CompositingReasons logic
for trace viewer and debug info. After this cleanup, the only
file someone needs to modify to add/remove compositing reasons
is CompositingReasons.h. All other uses of CompositingReasons
are for actual core logic.

This patch does the following:
1. Moves the CompositingReasons description strings
   into CompositingReasons.h,
2. Removes the redundant and out-dated CompositingReasons inspector
   layer tree agent, and instead adds necessary info to
   CompositingReasons.h,
3. Cleans up the compositing reason constants, and
4. Adds a few compositing reasons to the combo reasons,
   which seemed to be missing.

Review URL: https://codereview.chromium.org/171923002

git-svn-id: svn://svn.chromium.org/blink/trunk@167859 bbb929c8-8fbe-4397-9dbb-9b2b20218538