SerializedScriptValue: Don't hold isolate in Writer
The Writer (which mostly serializes C++ values to a buffer) does not
need the isolate. The Writer's isolate pointer was only used by the
Serializer which called into the Writer, and the Serializer already
has an isolate pointer.

This lets us remove a v8::Isolate::GetCurrent() call.

R=haraken
BUG=339838

Review URL: https://codereview.chromium.org/182183002

git-svn-id: svn://svn.chromium.org/blink/trunk@167978 bbb929c8-8fbe-4397-9dbb-9b2b20218538