Oilpan: Move StyleSheets to the oilpan heap using RefCountedGarbageCollected.
Resolve CSSStyleSheet/StyleSheetContents finalization order issue
by making the client pointers in StyleSheetContents weak. That
involves going to a hash set instead of a vector and therefore
we need a hash function for RawPtr.

The style engine change is in preparation for tracing the contained
collections of CSSStyleSheet pointers.

R=haraken@chromium.org, jochen@chromium.org, vegorov@chromium.org
BUG=

Review URL: https://codereview.chromium.org/173783002

git-svn-id: svn://svn.chromium.org/blink/trunk@168030 bbb929c8-8fbe-4397-9dbb-9b2b20218538