Increase width of viewportConstrainedNotCompositingReasons
The member was no longer big enough to store all the 
enumerators so the setter could fail. The net result was
that if you tried to set the reason to 4, it *always*
looked like this reason changed. I was seeing a full
tree build on every mouse move!

R=abarth@chromium.org
BUG=None

Review URL: https://codereview.chromium.org/177653010

git-svn-id: svn://svn.chromium.org/blink/trunk@168181 bbb929c8-8fbe-4397-9dbb-9b2b20218538