Avoid double attribute lookup in HTMLElement::matchesReadWritePseudoClass()
Avoid double attribute lookup in HTMLElement::matchesReadWritePseudoClass().
Instead of calling fastHasAttribute() then fastGetAttribute(), call
fastGetAttribute() directly and check if the returned value is nullAtom.

fastHasAttribute() and fastGetAttribute() do exactly the same thing internally
and the only difference is the return type. The attribute lookup is a linear
search and thus it is a good idea to avoid doing it twice.

R=adamk, eseidel

Review URL: https://codereview.chromium.org/184613003

git-svn-id: svn://svn.chromium.org/blink/trunk@168318 bbb929c8-8fbe-4397-9dbb-9b2b20218538